% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{RaggedArray-package}
\alias{RaggedArray-package}
\title{Fast, In-place Manipulation of Ragged Arrays}
\description{
R/C++ implementation of Ragged Arrays using \pkg{Rcpp} modules.
Ragged array data is stored as a
memory-contiguous matrix, modified in-place, and grown as needed.
}
\seealso{
Other Ragged.Array.Docs: \code{\link{RaggedArrayNew}}
}

